Boolean description
Great! You just stored a number in a variable. Numbers are one data
type we use in programming. A second data type is called a boolean.

A boolean is like a light switch. It can only have two values. Just
like a light switch can only be on or off, a boolean can only be True
or False.