Simple math - using to the power of **
All that math can be done on a calculator, so why use Python? Because
you can combine math with other data types (e.g. booleans) and commands
to create useful programs. Calculators just stick to numbers.

Now let's work with exponents.

eight = 2 ** 3
In the above example, we create a new variable called eight and set it
to 8, or the result of 2 to the power to 3 (2^3).

Notice that we use ** instead of * or the multiplication operator.