Adding in tipping percentage
You received good service, so you'd like to leave a 15% tip on top of
the cost of the meal, including tax.

Before we compute the tip for your bill, let's set a variable for the
tip. Again, we need to get the decimal form of the tip, so we divide
15.0 by 100.