Tipping calc - reassigning a variable
We've got the three variables we need to perform our calculation, and
we know some arithmetic operators that can help us out.

We saw in Lesson 1 that we can reassign variables. For example, we
could say spam = 7, then later change our minds and say spam = 3.
reassign meal to the value of itself + itself * tax. And yes, you're
allowed to reassign a variable in terms of itself!