Tipping calculator part 4
Now that meal has the cost of the food plus tax, let's introduce on
line 8 a new variable, total, equal to the new meal + meal * tip.

The code on line 10 formats and prints to the console the value of
total with exactly two numbers after the decimal. (We'll learn about
string formatting, the console, and print in Unit 2!)