String formatting
Remember, we used the % operator to replace the %s placeholders with
the variables in parentheses.

name = "Mike"
print "Hello %s" % (name)
You need the same number of %s terms in a string as the number of
variables in parentheses:

print "The %s who %s %s!" % ("Knights", "say", "Ni")
# This will print "The Knights who say Ni!"
Instructions
Now it's your turn! We have ___ in the code to show you what you need
to change!

Inside the string, replace the three ___ with %s.
After the string but before the three variables, replace the final ___
with a %.
Hit Save & Submit Code.
Answer the questions in the console as they pop up! Type in your answer
and hit Enter.
?
Stuck? Get a hint!
Get live, fast support from Codecademy Advisors when you get stuck.
Learn more.
Remember, the syntax is:

print "%s" % (string_variable)
Are you stuck with a spinning gear and it feels like your code is
frozen? That's because the code is asking you three questions! Click
inside the console window, type in your name, and then press Enter (or
Return).

The \ character on line 5 is a continuation marker. It simply tells
Python that line 5 continues onto line 6.