Comparators ==, !=, >, < etc
Let's start with the simplest aspect of control flow: comparators.
There are six:

Equal to (==)
Not equal to (!=)
Less than (<)
Less than or equal to (<=)
Greater than (>)
Greater than or equal to (>=)
Comparators check if a value is (or is not) equal to, greater than (or
equal to), or less than (or equal to) another value.

Note that == compares whether two things are equal, and = assigns a
value to a variable.