Boolean operators theory
Boolean operators compare statements and result in boolean values.
There are three boolean operators:

and, which checks if both the statements are True;
or, which checks if at least one of the statements is True;
not, which gives the opposite of the statement.
We'll go through the operators one by one.