Pyg latin - variable
Now we can get ready to start translating to Pig Latin! Let's review
the rules for translation:

You move the first letter of the word to the end and then append the
suffix 'ay'.
Example: python -> ythonpay

Let's create a variable to hold our translation suffix.

Instructions
Create a variable named pyg and set it equal to the suffix 'ay'.