Functions 9 import
Great! We've found a way to handpick the variables and functions we
want from modules.

What if we still want all of the variables and functions in a module
but don't want to have to constantly type math.?

Universal import can handle this for you. The syntax for this is:

from module import *
Instructions
Use the power of from module import * to import everything from the
math module on line 3 of the editor.