Lists
Lists are a datatype you can use to store a collection of different
pieces of information as a sequence under a single variable name.
(Datatypes you've already learned about include strings, numbers, and
booleans.)

You can assign items to a list with an expression of the form

list_name = [item_1, item_2]
with the items in between brackets. A list can also be empty:
empty_list = [].

Lists are very similar to strings, but there are a few key differences.

Instructions
The list zoo_animals has three items (check them out on line 1). Go
ahead and add a fourth! Just enter the name of your favorite animal (as
a "string") on line 1, after the final comma but before the closing ].